LayoutTests/imported/w3c/web-platform-tests/streams/piping/error-propagation-forward-expected.txt

CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream
CONSOLE MESSAGE: line 12: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: line 9: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: cancel() called on a reader owned by no readable stream

Harness Error (FAIL), message = cancel() called on a reader owned by no readable stream

PASS Errors must be propagated forward: starts errored; preventAbort = false; fulfilled abort promise 
FAIL Errors must be propagated forward: starts errored; preventAbort = false; rejected abort promise assert_throws: pipeTo must reject with the abort error function "function () { throw e }" threw object "error1: error1!" ("error1") expected object "error2: error2!" ("error2")
PASS Errors must be propagated forward: starts errored; preventAbort = undefined (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = null (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = false (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = 0 (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = -0 (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = NaN (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort =  (falsy); fulfilled abort promise 
PASS Errors must be propagated forward: starts errored; preventAbort = true (truthy) 
PASS Errors must be propagated forward: starts errored; preventAbort = a (truthy) 
PASS Errors must be propagated forward: starts errored; preventAbort = 1 (truthy) 
PASS Errors must be propagated forward: starts errored; preventAbort = Symbol() (truthy) 
PASS Errors must be propagated forward: starts errored; preventAbort = [object Object] (truthy) 
PASS Errors must be propagated forward: starts errored; preventAbort = true, preventCancel = true 
PASS Errors must be propagated forward: starts errored; preventAbort = true, preventCancel = true, preventClose = true 
PASS Errors must be propagated forward: becomes errored while empty; preventAbort = false; fulfilled abort promise 
FAIL Errors must be propagated forward: becomes errored while empty; preventAbort = false; rejected abort promise assert_throws: pipeTo must reject with the abort error function "function () { throw e }" threw object "error1: error1!" ("error1") expected object "error2: error2!" ("error2")
PASS Errors must be propagated forward: becomes errored while empty; preventAbort = true 
PASS Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = false; fulfilled abort promise 
FAIL Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = false; rejected abort promise assert_throws: pipeTo must reject with the abort error function "function () { throw e }" threw object "error1: error1!" ("error1") expected object "error2: error2!" ("error2")
PASS Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = true 
PASS Errors must be propagated forward: becomes errored after one chunk; preventAbort = false; fulfilled abort promise 
FAIL Errors must be propagated forward: becomes errored after one chunk; preventAbort = false; rejected abort promise assert_throws: pipeTo must reject with the abort error function "function () { throw e }" threw object "error1: error1!" ("error1") expected object "error2: error2!" ("error2")
PASS Errors must be propagated forward: becomes errored after one chunk; preventAbort = true 
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = false; fulfilled abort promise assert_array_equals: lengths differ, expected 2 got 4
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = false; rejected abort promise assert_throws: pipeTo must reject with the abort error function "function () { throw e }" threw object "error1: error1!" ("error1") expected object "error2: error2!" ("error2")
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = true assert_array_equals: lengths differ, expected 0 got 2
FAIL Errors must be propagated forward: shutdown must not occur until the final write completes assert_array_equals: lengths differ, expected 2 got 4
